requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
ollama
pypdf